+++ components/chat/ChatBubble.tsx
